# w4/openapi-comparison/2_APIBlueprint/out/python-server/openapi_server/models/book.py
from datetime import date, datetime  # noqa: F401

from typing import List, Dict  # noqa: F401

from openapi_server.models.base_model import Model
from openapi_server import util


class Book(Model):
    """NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).

    Do not edit the class manually.
    """

    def __init__(self, title=None, author=None, year=None, id=None):  # noqa: E501
        """Book - a model defined in OpenAPI

        :param title: The title of this Book.  # noqa: E501
        :type title: str
        :param author: The author of this Book.  # noqa: E501
        :type author: str
        :param year: The year of this Book.  # noqa: E501
        :type year: float
        :param id: The id of this Book.  # noqa: E501
        :type id: float
        """
        self.openapi_types = {
            'title': str,
            'author': str,
            'year': float,
            'id': float
        }

        self.attribute_map = {
            'title': 'title',
            'author': 'author',
            'year': 'year',
            'id': 'id'
        }

        self._title = title
        self._author = author
        self._year = year
        self._id = id

    @classmethod
    def from_dict(cls, dikt) -> 'Book':
        """Returns the dict as a model

        :param dikt: A dict.
        :type: dict
        :return: The Book of this Book.  # noqa: E501
        :rtype: Book
        """
        return util.deserialize_model(dikt, cls)

    @property
    def title(self) -> str:
        """Gets the title of this Book.


        :return: The title of this Book.
        :rtype: str
        """
        return self._title

    @title.setter
    def title(self, title: str):
        """Sets the title of this Book.


        :param title: The title of this Book.
        :type title: str
        """
        if title is None:
            raise ValueError("Invalid value for `title`, must not be `None`")  # noqa: E501

        self._title = title

    @property
    def author(self) -> str:
        """Gets the author of this Book.


        :return: The author of this Book.
        :rtype: str
        """
        return self._author

    @author.setter
    def author(self, author: str):
        """Sets the author of this Book.


        :param author: The author of this Book.
        :type author: str
        """

        self._author = author

    @property
    def year(self) -> float:
        """Gets the year of this Book.


        :return: The year of this Book.
        :rtype: float
        """
        return self._year

    @year.setter
    def year(self, year: float):
        """Sets the year of this Book.


        :param year: The year of this Book.
        :type year: float
        """

        self._year = year

    @property
    def id(self) -> float:
        """Gets the id of this Book.


        :return: The id of this Book.
        :rtype: float
        """
        return self._id

    @id.setter
    def id(self, id: float):
        """Sets the id of this Book.


        :param id: The id of this Book.
        :type id: float
        """

        self._id = id