# w4/openapi-comparison/2_APIBlueprint/out/python-server/openapi_server/models/book_base.py
from datetime import date, datetime  # noqa: F401

from typing import List, Dict  # noqa: F401

from openapi_server.models.base_model import Model
from openapi_server import util


class BookBase(Model):
    """NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).

    Do not edit the class manually.
    """

    def __init__(self, title=None, author=None, year=None):  # noqa: E501
        """BookBase - a model defined in OpenAPI

        :param title: The title of this BookBase.  # noqa: E501
        :type title: str
        :param author: The author of this BookBase.  # noqa: E501
        :type author: str
        :param year: The year of this BookBase.  # noqa: E501
        :type year: float
        """
        self.openapi_types = {
            'title': str,
            'author': str,
            'year': float
        }

        self.attribute_map = {
            'title': 'title',
            'author': 'author',
            'year': 'year'
        }

        self._title = title
        self._author = author
        self._year = year

    @classmethod
    def from_dict(cls, dikt) -> 'BookBase':
        """Returns the dict as a model

        :param dikt: A dict.
        :type: dict
        :return: The BookBase of this BookBase.  # noqa: E501
        :rtype: BookBase
        """
        return util.deserialize_model(dikt, cls)

    @property
    def title(self) -> str:
        """Gets the title of this BookBase.


        :return: The title of this BookBase.
        :rtype: str
        """
        return self._title

    @title.setter
    def title(self, title: str):
        """Sets the title of this BookBase.


        :param title: The title of this BookBase.
        :type title: str
        """
        if title is None:
            raise ValueError("Invalid value for `title`, must not be `None`")  # noqa: E501

        self._title = title

    @property
    def author(self) -> str:
        """Gets the author of this BookBase.


        :return: The author of this BookBase.
        :rtype: str
        """
        return self._author

    @author.setter
    def author(self, author: str):
        """Sets the author of this BookBase.


        :param author: The author of this BookBase.
        :type author: str
        """

        self._author = author

    @property
    def year(self) -> float:
        """Gets the year of this BookBase.


        :return: The year of this BookBase.
        :rtype: float
        """
        return self._year

    @year.setter
    def year(self, year: float):
        """Sets the year of this BookBase.


        :param year: The year of this BookBase.
        :type year: float
        """

        self._year = year